requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
djangorestframework>=2.3.12
simplejson>=3.3.3
coverage>=3.7.1
django-nose>=1.2
dougrain>=0.5.1
mock==1.0.1